Epoch data from the International Commission on Stratigraphy (v2022/10)

A dataset containing the boundary ages, abbreviations, and colors for the epochs of the Geologic Time Scale. Based on The ICS International Chronostratigraphic Chart (v2022/10), by Cohen, Finney, Gibbard, and Fan.

Usage

epochs

Format

A data frame with 34 rows and 5 variables:

name

epoch name

max_age

maximum age, in millions of years

min_age

minimum age, in millions of years

abbr

epoch name abbreviations

color

the colors for each epoch, according to the Commission for the Geological Map of the World

Source

https://stratigraphy.org via https://macrostrat.org/api/v2/defs/intervals?timescale=international%20epochs

See also