Pattern numbers and names for patterns from the FGDC Digital Cartographic Standard for Geologic Map Symbolization

A dataset containing the FGDC pattern numbers and verbatim names from the FGDC Digital Cartographic Standard for Geologic Map Symbolization by the U.S. Geological Survey and the Geologic Data Subcommittee (GDS) of the Federal Geographic Data Committee (FGDC). Only patterns with names are included (i.e., patterns 601-733).

Usage

fgdc_names

Format

A data frame with 117 rows and 2 variables:

code

the pattern number according to the FGDC standard

name

the verbatim name for the pattern according to the FGDC standard

Source

https://ngmdb.usgs.gov/fgdc_gds/geolsymstd/fgdc-geolsym-patternchart.pdf via https://github.com/davenquinn/geologic-patterns/

Details

These pattern numbers and names were originally extracted by Daven Quinn and are hosted on GitHub.

See also